[Silver III] Title: 주유소, Time: 376 ms, Memory: 35948 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -10,19 +10,19 @@ public class Main {
         int N = Integer.parseInt(br.readLine());
 
         StringTokenizer st_D = new StringTokenizer(br.readLine());
-        int[] D = new int[N - 1];
+        long[] D = new long[N - 1];
         for (int i = 0; i < N - 1; i++) {
             D[i] = Integer.parseInt(st_D.nextToken());
         }
 
         StringTokenizer st_P = new StringTokenizer(br.readLine());
-        int[] P = new int[N];
+        long[] P = new long[N];
         for (int i = 0; i < N; i++) {
             P[i] = Integer.parseInt(st_P.nextToken());
         }
 
-        int oil = 0;
-        int min = P[0];
+        long oil = 0;
+        long min = P[0];
         for (int i = 0; i < N - 1; i++) {
             if (P[i] < min) {
                 min = P[i];
